bugfix #157289, Summary.getTotalLiveBytes() Summary.getTotalLiveInstances() and JavaClass.getAllInstancesSize() now return long instead of int

+++ profiler/src/org/netbeans/modules/profiler/heapwalk/ClassesListController.java
@@ -46,7 +46,6 @@ import org.netbeans.api.project.Project;
 import org.netbeans.lib.profiler.global.CommonConstants;
 import org.netbeans.lib.profiler.heap.*;
 import org.netbeans.modules.profiler.heapwalk.ui.ClassesListControllerUI;
-import org.netbeans.modules.profiler.utils.SourceUtils;
 import org.openide.util.NbBundle;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -57,6 +56,7 @@ import java.util.Iterator;
 import java.util.List;
 import javax.swing.AbstractButton;
 import javax.swing.JPanel;
+import org.netbeans.modules.profiler.projectsupport.utilities.SourceUtils;
 
 
 /**
@@ -92,7 +92,7 @@ public class ClassesListController extends AbstractController {
                 case 2:
                     return new Integer(jClass1.getInstancesCount()).compareTo(new Integer(jClass2.getInstancesCount()));
                 case 3:
-                    return new Integer(jClass1.getAllInstancesSize()).compareTo(jClass2.getAllInstancesSize());
+                    return new Long(jClass1.getAllInstancesSize()).compareTo(jClass2.getAllInstancesSize());
                 default:
                     throw new RuntimeException("Unsupported compare operation for " + o1 + ", " + o2); // NOI18N
             }
